backend/requirements.txt
requests==2.26.0
Django==3.2
django-filter==22.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
Djoser==2.1.0
Pillow==9.5.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.21.1
gunicorn==20.0.4
psycopg2-binary==2.9.5
django_debug_toolbar==3.0.0
tzdata==2023.3
django-cors-headers==3.14.0